# server/swagger.yml
openapi: 3.1.0
info:
  title: Centralized Video Platform API
  description: |
    Learn how to interact with this API with your own client app.
    # How to Authenticate
    ## 1. Sign In Via Client App
    You can sign in on a client app which will place a cookie in your browser giving you access to protected routes.
    ## 2. Sign In With Swagger
    You can also use the **/auth/api/login** path below and provide your username and password in the request body. The response will include a token you can copy/paste into the Authorize button. Alternately, you can use an API tool like Postman and apply the token in the Authorization header with the value ```bearer [token]```
  version: 1.0.0
servers:
  - url: https://cvp.programminghelp.org/api/v1
  - url: http://localhost:3000/api/v1
    description: Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    successType:
      type: string
      enum:
        - success
        - error
...
    updateVideo:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
...
    success:
      properties:
        response:
          $ref: '#/components/schemas/successType'
        status:
          type: number
          examples:
            - 200
        message:
          type: string
          examples:
            - Successfully completed action.
        prevPath:
          type: string
          examples:
            - /
...
    unprocessableEntity:
      properties:
        response:
          $ref: '#/components/schemas/successType'
          examples:
            - error
        status:
          type: number
          examples:
            - 422
        message:
          type: string
          examples:
            - Invalid arguments.
        prevPath:
          type: string
          examples:
            - /
    internalServerError:
      properties:
        response:
          $ref: '#/components/schemas/successType'
          examples:
            - error
        status:
          type: number
          examples:
            - 500
        message:
          type: string
          examples:
            - Something went wrong.
        prevPath:
          type: string
          examples:
            - /
...
paths:
...
  /video/{username}/{video}:
    put:
      description: Update a video.
      tags:
        - Videos
      security:
        - bearerAuth: []
      requestBody:
        description: YouTube key or full URL of video to add.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateVideo'
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: Username found in path of URL.
        - in: path
          name: video
          required: true
          schema:
            type: string
          description: Video id
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/success'
        '401':
          description: Unauthenticated
        '403':
          description: Unauthorized
        '404':
          description: User or video not found.
        '422':
          description: >-
            Unprocessable Content - Occurs when there is an error with the
            username or video ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/unprocessableEntity'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalServerError'
    delete:
      description: Delete a video.
      tags:
        - Videos
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: Username found in path of URL.
        - in: path
          name: video
          required: true
          schema:
            type: string
          description: Video id
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/success'
        '401':
          description: Unauthenticated
        '403':
          description: Unauthorized
        '404':
          description: User or video not found.
        '422':
          description: >-
            Unprocessable Content - Occurs when there is an error with the
            username or video ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/unprocessableEntity'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalServerError'
...